+++ src/main/java/firok/tiths/util/CompoHandle.java
@@ -9,6 +9,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 public @interface CompoHandle
 {
-	float modifier();
+	double modifier();
 	int durability();
 }
